SignUpTatianaTTest added into group_eat_and_drink_Java
